Move scope-dialog files the scope-dialog directory
