+++ src/app/app.module.ts
@@ -22,7 +22,7 @@ import { ResponseStatusBarComponent } from "./response-status-bar.component";
 import { SampleCategoriesPanelComponent } from "./sample-categories-panel.component";
 import { RequestEditorsComponent } from "./request-editors.component";
 import { ShareLinkBtnComponent } from "./share-link-btn.component";
-import { ScopesDialogComponent } from "./scopes-dialog.component";
+import { ScopesDialogComponent } from "./scopes-dialog/scopes-dialog.component";
 import { GenericDialogComponent } from "./generic-message-dialog.component";
 
 @NgModule({
