+++ src/app/scopes-dialog/scopes-dialog.component.css
